move all database related code into separate directory
