+++ Server/db/db_character_inventory_tbl.h
