+++ Server/db/db_character_type_tbl.h
