+++ Server/main.c
@@ -49,18 +49,18 @@ To compile server, link with the following libraries :
 #include "server_messaging.h"
 #include "clients.h"
 #include "game_data.h"
-#include "database_functions.h"
+#include "db/database_functions.h"
 #include "client_protocol_handler.h"
 #include "server_protocol_functions.h"
-#include "db_character_tbl.h"
-#include "db_character_race_tbl.h"
-#include "db_map_tbl.h"
-#include "db_character_type_tbl.h"
-#include "db_gender_tbl.h"
-#include "db_chat_channel_tbl.h"
-#include "db_game_data_tbl.h"
-#include "db_attribute_tbl.h"
-#include "db_season_tbl.h"
+#include "db/db_character_tbl.h"
+#include "db/db_character_race_tbl.h"
+#include "db/db_map_tbl.h"
+#include "db/db_character_type_tbl.h"
+#include "db/db_gender_tbl.h"
+#include "db/db_chat_channel_tbl.h"
+#include "db/db_game_data_tbl.h"
+#include "db/db_attribute_tbl.h"
+#include "db/db_season_tbl.h"
 #include "date_time_functions.h"
 #include "broadcast_actor_functions.h"
 #include "movement.h"
